requirements.txt
# 遥感数据集生成器依赖包
# Remote Sensing Dataset Generator Dependencies

# 核心地理数据处理
geopandas>=0.13.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.5.0

# 图像处理
Pillow>=9.5.0
numpy>=1.24.0

# HTTP请求和异步处理
requests>=2.31.0
aiohttp>=3.8.0
aiofiles>=23.0.0

# 数据序列化和配置
pyyaml>=6.0
toml>=0.10.2

# 进度条和用户界面
tqdm>=4.65.0
rich>=13.0.0

# 系统监控和性能
psutil>=5.9.0

# 日志和调试
loguru>=0.7.0

# 开发和测试依赖（可选）
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0